+++ crates/bevy_auto_plugin_shared/src/syntax/validated/non_empty_path.rs
@@ -29,7 +29,7 @@ impl NonEmptyPath {
         Self(path)
     }
     pub fn into_last_ident(self) -> Ident {
-        self.0.segments.into_iter().last().expect("non-empty path").ident
+        self.0.segments.into_iter().next_back().expect("non-empty path").ident
     }
     pub fn last_ident(&self) -> &Ident {
         &self.0.segments.last().expect("non-empty path").ident